bpf: lift LRP code into its own header filer
Make it a bit easier to identify the LRP-specific pieces.
