+++ bpf/bpf_alignchecker.c
@@ -108,5 +108,6 @@ add_type(struct mcast_subscriber_v4);
 add_type(struct node_key);
 add_type(struct node_value);
 
+#include "lib/lrp.h"
 add_type(struct skip_lb4_key);
 add_type(struct skip_lb6_key);
